+++ penman/transform.py
@@ -4,6 +4,7 @@ Tree and graph transformations.
 """
 
 from typing import Union, List, Tuple
+import logging
 
 from penman.types import BasicTriple
 from penman.epigraph import (Epidatum, Epidata)
@@ -12,6 +13,9 @@ from penman.tree import (Tree, Node, is_atomic)
 from penman.graph import (Graph, CONCEPT_ROLE)
 from penman.model import Model
 from penman.layout import (Push, POP)
+
+
+logger = logging.getLogger(__name__)
 
 
 def canonicalize_roles(t: Tree, model: Model) -> Tree:
@@ -40,7 +44,9 @@ def canonicalize_roles(t: Tree, model: Model) -> Tree:
     """
     if model is None:
         model = Model()
-    return Tree(_canonicalize_node(t.node, model), metadata=t.metadata)
+    tree = Tree(_canonicalize_node(t.node, model), metadata=t.metadata)
+    logger.info('Canonicalized roles: %s', tree)
+    return tree
 
 
 def _canonicalize_node(node: Node, model: Model) -> Node:
@@ -96,9 +102,11 @@ def reify_edges(g: Graph, model: Model) -> Graph:
             # the tree; maybe this should be a tree operation?
         else:
             new_triples.append(triple)
-    return Graph(new_triples,
-                 epidata=new_epidata,
-                 metadata=g.metadata)
+    g = Graph(new_triples,
+              epidata=new_epidata,
+              metadata=g.metadata)
+    logger.info('Reified edges: %s', g)
+    return g
 
 
 def contract_edges(g: Graph, model: Model) -> None:
@@ -151,9 +159,11 @@ def reify_attributes(g: Graph) -> Graph:
             new_epidata[node_triple] = node_epis + [POP]
         else:
             new_triples.append(triple)
-    return Graph(new_triples,
-                 epidata=new_epidata,
-                 metadata=g.metadata)
+    g = Graph(new_triples,
+              epidata=new_epidata,
+              metadata=g.metadata)
+    logger.info('Reified attributes: %s', g)
+    return g
 
 
 def indicate_branches(g: Graph, model: Model) -> Graph:
@@ -203,9 +213,11 @@ def indicate_branches(g: Graph, model: Model) -> Graph:
                 assert isinstance(t[2], str)
                 new_triples.append((t[2], model.top_role, t[0]))
         new_triples.append(t)
-    return Graph(new_triples,
-                 epidata=g.epidata,
-                 metadata=g.metadata)
+    g = Graph(new_triples,
+              epidata=g.epidata,
+              metadata=g.metadata)
+    logger.info('Indicated branches: %s', g)
+    return g
 
 
 _SplitMarkers = Tuple[Union[Epidatum, None], Epidata, Epidata, Epidata]